+++ src/main/java/com/github/pagehelper/PageHelper.java
@@ -60,17 +60,6 @@ public class PageHelper implements Interceptor {
      *
      * @param pageNum  页码
      * @param pageSize 每页显示数量
-     * @param orderBy  针对sqlserver - 建议在sql中直接包含order by
-     */
-    public static Page startPage(int pageNum, int pageSize, String orderBy) {
-        return startPage(pageNum, pageSize, true).orderBy(orderBy);
-    }
-
-    /**
-     * 开始分页
-     *
-     * @param pageNum  页码
-     * @param pageSize 每页显示数量
      * @param count    是否进行count查询
      */
     public static Page startPage(int pageNum, int pageSize, boolean count) {